requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==0.61.1
djangorestframework==3.15.2
idna==3.10
ngrok==1.4.0
oauthlib==3.2.2
pillow==11.0.0
psycopg==3.2.3
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyngrok==7.2.1
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.1
